requirements.txt
beautifulsoup4~=4.13.3
certifi~=2024.2.2
charset-normalizer~=3.4.1
httplib2~=0.22.0
idna~=3.10
oauth2client~=4.1.3
pip~=24.3.1
pyasn1~=0.6.1
pyasn1_modules~=0.4.2
pyparsing~=3.2.3
PyYAML~=6.0.2
requests~=2.32.3
rsa~=4.9
six~=1.17.0
soupsieve~=2.6
typing_extensions~=4.13.1
urllib3~=2.2.3
